Add files via upload
These are finals. for the actions. The Directories will need to be modified for your COMPLETE directory whether that be a TOR/NEWSGROUP/DVR

Modify the **.ts/mkv to match what videos you may download.

In my case the DVR records in TS and i am transcoding it to MKV with a comskip action as well.